+++ src/types/pomodoro.ts
@@ -4,6 +4,14 @@ export type PomodoroSettings = {
   primaryColor: string;
   secondaryColor: string;
   tertiaryColor: string;
+  autoPlayRadioOnPomodoroStart: boolean;
+  lowerRadioVolumeOnBreak: boolean;
+};
+
+export type RadioPlayerState = {
+  isMuted: boolean;
+  previousVolume: number;
+  volume: number;
 };
 
 export type PomodoroState = {